+++ package.json
@@ -30,7 +30,6 @@
   },
   "devDependencies": {
     "jasmine": "^3.5.0",
-    "nock": "^11.7.0",
     "nyc": "^14.1.1",
     "request": "^2.88.0",
     "request-promise": "^4.2.4"
