+++ spec/Integration/utils/Utils.js
@@ -1,7 +1,5 @@
 const rp = require('request-promise')
-const nock = require('nock')
 
-const mockApi = nock('http://localhost:8081')
 const BASE_URL = 'http://localhost:8081'
 
 function apiCall(obj) {
@@ -17,14 +15,4 @@ function apiCall(obj) {
   return rp(options)
 }
 
-function generateRandomString() {
-  return (
-    Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15)
-  )
-}
-
-module.exports = {
-  apiCall,
-  mockApi,
-  generateRandomString
-}
+module.exports = apiCall
